+++ lib/smb2-cmd-negotiate.c
@@ -277,7 +277,7 @@ smb2_process_negotiate_fixed(struct smb2_context *smb2,
         smb2_get_uint16(iov, 58, &rep->security_buffer_length);
 
         if (rep->security_buffer_length &&
-            (rep->security_buffer_offset + rep->security_buffer_length > smb2->spl)) {
+            (rep->security_buffer_offset + rep->security_buffer_length > (uint16_t)smb2->spl)) {
                 smb2_set_error(smb2, "Security buffer extends beyond end of "
                                "PDU");
                 return -1;
